# docs/reference/api/openapi-v1.yaml
openapi: 3.0.3
info:
  title: OpenManager AI API (v1)
  version: 2026.02.28
  contact:
    name: OpenManager Team
    email: support@openmanager-ai.vercel.app
  description: |
    Contract-first API baseline for high-complexity endpoints.
    Scope excludes complete product surface; remaining routes are in the route catalog.
servers:
  - url: /
    description: Vercel API base
paths:
...
  /api/ai/jobs:
    get:
      operationId: listJobs
      summary: List AI jobs for a session
      security:
        - sessionCookie: []
        - secureSessionCookie: []
        - guestSessionCookie: []
        - apiKeyAuth: []
        - testSecretAuth: []
      parameters:
        - in: query
          name: sessionId
          required: true
          schema:
            type: string
          description: Job list session scope
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Job list
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                type: object
                required: [jobs, total, hasMore]
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/JobStatusResponse'
                  total:
                    type: integer
                  hasMore:
                    type: boolean
        '400':
          description: Invalid sessionId or request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
    post:
      operationId: createJob
      summary: Create AI job
      security:
        - sessionCookie: []
        - secureSessionCookie: []
        - guestSessionCookie: []
        - apiKeyAuth: []
        - testSecretAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                options:
                  type: object
                  required: []
                  properties:
                    sessionId:
                      type: string
                    timeoutMs:
                      type: integer
                      minimum: 100
                type:
                  type: string
      responses:
        '201':
          description: Job accepted
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
            X-RateLimit-Daily-Limit:
              $ref: '#/components/headers/X-RateLimit-Daily-Limit'
            X-RateLimit-Daily-Remaining:
              $ref: '#/components/headers/X-RateLimit-Daily-Remaining'
            X-RateLimit-Daily-Reset:
              $ref: '#/components/headers/X-RateLimit-Daily-Reset'
          content:
            application/json:
              schema:
                type: object
                required: [jobId, status, pollUrl, estimatedTime, triggerStatus]
                properties:
                  jobId:
                    type: string
                  status:
                    type: string
                  pollUrl:
                    type: string
                    format: uri
                  estimatedTime:
                    type: number
                    minimum: 0
                  triggerStatus:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Redis unavailable or queue worker unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
            X-RateLimit-Daily-Limit:
              $ref: '#/components/headers/X-RateLimit-Daily-Limit'
            X-RateLimit-Daily-Remaining:
              $ref: '#/components/headers/X-RateLimit-Daily-Remaining'
            X-RateLimit-Daily-Reset:
              $ref: '#/components/headers/X-RateLimit-Daily-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
  /api/ai/jobs/{id}:
    get:
      operationId: getJobById
      summary: Get AI job status by id
      security:
        - sessionCookie: []
        - secureSessionCookie: []
        - guestSessionCookie: []
        - apiKeyAuth: []
        - testSecretAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job status
          headers:
            Cache-Control:
              schema:
                type: string
            X-Job-Status:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
        '400':
          description: Invalid job id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job not found or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: cancelJob
      summary: Cancel AI job by id
      security:
        - sessionCookie: []
        - secureSessionCookie: []
        - guestSessionCookie: []
        - apiKeyAuth: []
        - testSecretAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cancel success or idempotent success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  jobId:
                    type: string
                  status:
                    type: string
        '400':
          description: Invalid job id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    testSecretAuth:
      type: apiKey
      in: header
      name: x-test-secret
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
    secureSessionCookie:
      type: apiKey
      in: cookie
      name: __Secure-next-auth.session-token
    guestSessionCookie:
      type: apiKey
      in: cookie
      name: auth_session_id
  headers:
    X-RateLimit-Limit:
      schema:
        type: integer
        minimum: 0
    X-RateLimit-Remaining:
      schema:
        type: integer
        minimum: 0
    X-RateLimit-Reset:
      schema:
        type: string
        description: Epoch ms timestamp
    X-RateLimit-Daily-Limit:
      schema:
        type: integer
        minimum: 0
    X-RateLimit-Daily-Remaining:
      schema:
        type: integer
        minimum: 0
    X-RateLimit-Daily-Reset:
      schema:
        type: string
        description: Epoch ms timestamp
    Retry-After:
      schema:
        type: string
        description: Retry delay in seconds
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
        fallback:
          oneOf:
            - type: boolean
            - type: string
        isServerError:
          type: boolean
        notFound:
          type: boolean
        details:
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items:
                type: string
            - type: string
    RateLimitError:
      type: object
      required: [error, message, retryAfter]
      properties:
        error:
          type: string
          example: Too Many Requests
        message:
          type: string
        retryAfter:
          type: integer
          minimum: 0
        dailyLimitExceeded:
          type: boolean
    JobStatusResponse:
      type: object
      required: [jobId, type, status, progress, createdAt]
      properties:
        jobId:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [queued, processing, completed, failed, cancelled]
        progress:
          type: number
          minimum: 0
          maximum: 100
        currentStep:
          type: string
          nullable: true
        result:
          type: object
          nullable: true
          properties:
            content:
              type: string
        error:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          nullable: true
          format: date-time
        completedAt:
          type: string
          nullable: true
          format: date-time
...